Add the `com` module
Add the `com` module to EpiJS, which includes the community and virus classes. The community class allows for representation of a community, and the virus class represents a virus. These speed up the process in which you can create a model. You can use the same virus several times on a community.

There is also the compare function, which compares an outbreak of two different communities.

Commit overview:

  docs/.vuepress/config.js:
    - Update sidebar for com module.

  docs/com.md:
    - Update with docs for com module

  docs/pre.md:
    - Update with docs for pre module

  EpiJS/com.js:
    - Add the community class
    - Add the virus class
    - Add the compare function

  EpiJS/index.js:
    - Export the com module

  scripts/doc-update.py:
    - Update script for com module

  tests/js/pre.js:
    - Add tests for the community and virus classes.
    - Add tests for the compare function

  tests/index.html:
    - Add tests for the community and virus classes.
    - Add tests for the compare function

  web/com.min.js:
    - Update with minified JavaScript from EpiJS/com.js

  web/pre.min.js:
    - Update with minified JavaScript from EpiJS/pre.js

  package.json:
    - Update documentation script to generate for com module.