Fix deepsource errors (5)
Fix deepsource errors `JS-0119`, `JS-0045`, and `JS-0050`. Both
`JS-0119` and `JS-0045` were silenced. `JS-0045` states that a function
should always return somethin, and in the case of the get_data function
in model.js, it does, so it was silenced. `JS-0119` stated that variable
should be assigned during declaration, but since the variables in both
comp.js and model.js were used in a for/in statement, declaration was
not necassary, and hence it was silenced. However, issue `JS-0050` was
fixed, which stated that `==` and `!=` should not be used, and only
`===` or `!==` should be.

Commit Overview:
  EpiJS/comp.js:
    - Fix `JS-0050` issue.
    - Silence `JS-0119` issue.
  EpiJS/model.js:
    - Silence `JS-0119` issue.
    - Silence `JS-0045` issue.