[#37] Fix bugs in the `com` module
Fix two bugs for the stochastic models in the `com` module. The first in `pre`, which was letting values reach negative, and therefore in the next iteration the value would become NaN. Now, it resets it to 0 if it goes below 0. The other bug in the pre module was a formula error, which added the noise and functions for stochastic models after all the original equations were parsed, although this should be added when the specific set is run.

CHANGELIST:
  * Fix formula error
  * Fix bug which outputs NaN as value of a population
  * Update and fix bug in tests
  * Update `web/index.min.js` with new built code

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>