[#37] Add support for stochastic models in the `comp` module.
Add support for stochastic models in the `comp` module, which is defined in initialization. This allows for more flexibility in EpiJS, instead of manually passing in noise into the function. This also works in the Idiom class, however, EpiJS does not auto generate the equations, but will generate noise with the Gaussian distribution.

CHANGELIST:
  * Add tests for the new stochastic models in comp
  * Add stochastic model support in comp
  * Update build with latest functions

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>