Patch 37 (#56)
* [#37] Add stochastic modelling to `pre` module

Adds stochastic modelling to the `pre` module, and add tests and new docs for the new feature. This commit also introduces a new dependency gaussian, for creating random numbers with a normal distribution, which is used in stochastic modelling.

CHANGELIST:
  * Add stochastic modelling to all functions in the pre module
  * Add documentation for the new functions
  * Add tests for the new feature
  * Update minified code
  * Add new dependency "gaussian".
  * Fix a bug in two pre functions
  * Fix error in test script, which doesn't detect errors

Signed-off-by: Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Upgrade dependencies

Upgrade all dependencies, including those which are dependencies of dependencies and dev dependencies.

Signed-off-by: Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Fix docs workflow errors

* [#37] Add stochastic models to the `com` module

Adds stochastic models to the com module as detailed in #37. This adds the optional feature to the sir, seir, and seird functions of the Community class, and add new tests for the function.

CHANGELIST:
  * Add stochastic models to the sir, seir, and seird functions of the Community class of the com module
  * Add new tests for the stochastic models
  * Re-build the code for the stochastic models

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* [#37] Fix bugs in the `com` module

Fix two bugs for the stochastic models in the `com` module. The first in `pre`, which was letting values reach negative, and therefore in the next iteration the value would become NaN. Now, it resets it to 0 if it goes below 0. The other bug in the pre module was a formula error, which added the noise and functions for stochastic models after all the original equations were parsed, although this should be added when the specific set is run.

CHANGELIST:
  * Fix formula error
  * Fix bug which outputs NaN as value of a population
  * Update and fix bug in tests
  * Update `web/index.min.js` with new built code

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* [#37] Add support for stochastic models in the `comp` module.

Add support for stochastic models in the `comp` module, which is defined in initialization. This allows for more flexibility in EpiJS, instead of manually passing in noise into the function. This also works in the Idiom class, however, EpiJS does not auto generate the equations, but will generate noise with the Gaussian distribution.

CHANGELIST:
  * Add tests for the new stochastic models in comp
  * Add stochastic model support in comp
  * Update build with latest functions

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Fix comp.js bug introduced in c79c74e

Fix comp.js bug introduced in c79c74e which defined an undeclared variable without prefixing it with "let", "var", or "const".

Signed-off-by: Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Ignore and remove lock files

* Fix deepsource errors

Silence 5 JS-0123 errors, and fix 24 JS-0050 errors, 11 JS-0119 errors, and 113 JS-0067 errors.

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Fix 9 additional deepsource errors

Fix 9 additional deepsource errors of the JS-0119 error.

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Fix declaration error

Fix declaration error which didn't properly define variable omega1 in Community.sir, seir, and seird.

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>