[#59] Add `plots` module (#61)
* Make stochastic option required

Makes the stochastic parameter on the functions in com, comp, and pre required and no longer optional.

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Remove canvas output from pre module

Removes the canvas output from the pre module as detailed in #59. This will be one of the larger deprecations in the v2.0 release. This now instead outputs a list of objects each of which contain the population numbers of the compartments

CHANGELIST:
  * Fix stochastic model bug in comp
  * Return a list for all pre module functions instead of plotting the functions
  * Update tests with the new updates
  * Update build with the new updates

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Add CDNJS to installation docs

Add the new CDNJS integration to the installation page in the cdn category

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Successfully introduce the `plots` module

Introduce the `plots` module, which plots output from the `pre` module, and will be extended in the future to plot com, comp, and model modules outputs. This also officially fixes #59.

CHANGELIST:
  * Add the plots module
  * Fix spelling error in model module
  * Remove unnecassary import of chart.js in the pre module
  * Update tests with new functions
  * Add test for the plots module
  * Update funcs.js to include the plots module plot function
  * Re-build the code

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* Fix 2 deepsource errors

Fix 2 deepsource errors: JS-0059 and JS-0050.

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

---


Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>