[#70] Export Models as Files (#72)
* [#70] Add model import and export

Add model importing and exporting from json files. The feature is NodeJS only, as writing and reading files on the web is not possible, even with browserify.

CHANGELIST:
  * Add exporting of models to JSON files
  * Add importiing of models from JSON files
  * Update code with new web built code.

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

* [#70] Add exporting with custom file types

Allows users to define custom file types for exporting, currently only js and json are supported. This might be extended in the future to YAML.

CHANGELIST:
  * Add exporting for custom file types
  * Add JavaScript exporting for models
  * Add model importing with custom user-defined file types
  * Add JS parser for importing with .js files

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>

---

Signed-off-by: @Quantalabs <55121845+Quantalabs@users.noreply.github.com>